requirements.txt
# Core dependencies
pydantic>=2.5.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# SQLite (built-in Python)
# MSSQL (opzionale)
pyodbc>=4.0.0

# Networking
requests>=2.31.0
aiohttp>=3.9.0
aiohttp-cors>=0.7.0
websockets>=12.0
httpx>=0.25.0

# CLI
click>=8.1.0
rich>=13.7.0

# Utils
jinja2>=3.1.0
toml>=0.10.2

# Validation (opzionale)
email-validator>=2.1.0
validators>=0.22.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0

pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
loguru>=0.7.0
sqlalchemy>=2.0.0
requests>=2.31.0
click>=8.1.0
rich>=13.7.0